pyproject.toml
[tool.poetry]
name = "python-bigtube"
version = "0.1.0"
description = "Universal Multimedia Downloader using GTK4/Adwaita"
authors = ["eltonff <eltonfabricio10@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bigtube", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
PyGObject = "^3.44"
requests = "^2.31.0"
python-mpv = "^1.0.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.1"
mypy = "^1.8.0"

[tool.poetry.scripts]
bigtube = "bigtube.main:run"
